requirements.txt
pandas
numpy
matplotlib
seaborn
sqlalchemy
jupyter
ipykernel
python-dotenv
pyarrow
